arch: fix internal/fix → internal/engine import inversion (plan 204)
Move CheckRules and ConfigureRule to a new internal/checker package,
and DedupeDiagnostics to internal/lint, so internal/fix can call
shared helpers without importing internal/engine (which sits above
fix in the architecture layer map). Add an integration contract test
that fails if fix ever re-imports engine.
