+++ plan/204_arch-fix-fix-engine-inversion.md
@@ -1,7 +1,7 @@
 ---
 id: 204
 title: Fix internal/fix importing internal/engine
-status: "🔲"
+status: "✅"
 summary: >-
   internal/fix/fix.go imports internal/engine for
   CheckRules, ConfigureRule, and DedupeDiagnostics.
@@ -22,30 +22,44 @@ the three functions to a package both
 engine and fix can import restores the
 intended direction.
 
+## Target package choice
+
+- `DedupeDiagnostics` moved to `internal/lint`:
+  it only operates on `lint.Diagnostic` values and
+  has no other dependencies.
+- `CheckRules` and `ConfigureRule` moved to a new
+  `internal/checker` package: both need `config.RuleCfg`,
+  `rule.Rule`, and `lint.File`. The `rule` package
+  already imports `config` (which imports `rule`), so
+  adding `config` to `rule` would cycle. A new
+  `internal/checker` package sits below both
+  `internal/engine` and `internal/fix` with no
+  import of either.
+
 ## Tasks
 
-1. Read `CheckRules`, `ConfigureRule`,
+1. [x] Read `CheckRules`, `ConfigureRule`,
    and `DedupeDiagnostics` and identify
    their dependencies.
-2. Choose the target package (`internal/lint`,
+2. [x] Choose the target package (`internal/lint`,
    `internal/rule`, or a new package).
    Document the choice here.
-3. Move the three functions.
-4. Update all callers in `internal/engine`
+3. [x] Move the three functions.
+4. [x] Update all callers in `internal/engine`
    and `internal/fix`.
-5. Verify `internal/fix` no longer
+5. [x] Verify `internal/fix` no longer
    imports `internal/engine`.
-6. Add a contract test in
+6. [x] Add a contract test in
    `internal/integration/` that fails if
    `internal/fix` imports `internal/engine`.
-7. Run `go build ./...` and
+7. [x] Run `go build ./...` and
    `go test ./...`.
 
 ## Acceptance Criteria
 
-- [ ] `grep -r --include='*.go' '".*internal/engine"' internal/fix/`
+- [x] `grep -r --include='*.go' '".*internal/engine"' internal/fix/`
   returns nothing.
-- [ ] A contract test guards the boundary.
-- [ ] `go build ./...` clean.
-- [ ] `go test ./...` passes.
-- [ ] `go tool golangci-lint run` clean.
+- [x] A contract test guards the boundary.
+- [x] `go build ./...` clean.
+- [x] `go test ./...` passes.
+- [x] `go tool golangci-lint run` clean.
